requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
ultralytics==8.3.27
pillow==11.0
python-multipart==0.0.17